\name{AdvancedRlab5-package}
\alias{AdvancedRlab5-package}
\alias{AdvancedRlab5}
\docType{package}
\title{
\packageTitle{AdvancedRlab5}
}
\description{
\packageDescription{AdvancedRlab5}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AdvancedRlab5}
\packageIndices{AdvancedRlab5}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{AdvancedRlab5}

Maintainer: \packageMaintainer{AdvancedRlab5}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
