% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Generic wrapper function to get JSON data from URL address}
\usage{
get_data(URL, query)
}
\arguments{
\item{URL}{a URL address}

\item{query}{API query}
}
\value{
A list containing the received data
}
\description{
Generic wrapper function to get JSON data from URL address
}
\examples{
get_data("http://api.kolada.se/v2/municipality", query="title=Linköping")
}
