% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_figures}
\alias{get_figures}
\title{function to get emission data for given municipality}
\usage{
get_figures(muni_name)
}
\arguments{
\item{muni_name}{name of municipality}
}
\value{
data frame containing different emission data by year
}
\description{
function to get emission data for given municipality
}
\examples{
get_figures("Linköping")
}
