% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koladabindings.R
\name{get_kpi}
\alias{get_kpi}
\title{Get KPI}
\usage{
get_kpi(name)
}
\arguments{
\item{name}{The name of the kpi we want to access.}
}
\value{
Returns a data frame that contains for a specific KPI with its id and description.
}
\description{
Get KPI
}
