% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koladabindings.R
\name{get_kpi_groups}
\alias{get_kpi_groups}
\title{Get KPI groups}
\usage{
get_kpi_groups(name)
}
\arguments{
\item{name}{The name of the group we want to access.}
}
\value{
Returns a data frame that contains information about a specific kpi group.
}
\description{
Get KPI groups
}
