% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_muni_id}
\alias{get_muni_id}
\title{function to get municipality ID}
\usage{
get_muni_id(muni_name)
}
\arguments{
\item{muni_name}{name of municipality}
}
\value{
the municipality ID
}
\description{
function to get municipality ID
}
