% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koladabindings.R
\name{get_municipality}
\alias{get_municipality}
\title{Get municipalities}
\usage{
get_municipality(name)
}
\arguments{
\item{name}{The name of the municipality we want to access.}
}
\value{
Returns a list that contains all municipalities.
}
\description{
Get municipalities
}
