% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koladabindings.R
\name{get_municipality_groups}
\alias{get_municipality_groups}
\title{Get a municipality's organizational units}
\usage{
get_municipality_groups(name)
}
\arguments{
\item{name}{The name/code id of the municipality's organizational unit we want to access.}
}
\value{
Returns a list that contains all organizational units of a municipality.
}
\description{
Get a municipality's organizational units
}
