% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koladabindings.R
\name{kolada_api}
\alias{kolada_api}
\title{kolada_api handles the API link}
\usage{
kolada_api(path)
}
\arguments{
\item{path}{The path of the API url.}
}
\value{
Returns a list that contains all data for the given API.
}
\description{
kolada_api handles the API link
}
