% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{plot_figures}
\alias{plot_figures}
\title{function to plot emission data of given municipality}
\usage{
plot_figures(muni_name)
}
\arguments{
\item{muni_name}{name of municipality}
}
\value{
a list of ggplots
}
\description{
function to plot emission data of given municipality
}
\examples{
plot_figures("Linköping")
}
