% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveplot.R
\name{visualize_kolada}
\alias{visualize_kolada}
\title{Visualize Kolada
Starts Kolada visualization demo.}
\usage{
visualize_kolada()
}
\value{
Nothing.
}
\description{
Visualize Kolada
Starts Kolada visualization demo.
}
