% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koladabindings.R
\name{get_search_results}
\alias{get_search_results}
\title{Get Search Results}
\usage{
get_search_results(kpi_list = NULL, municipality_list = NULL, year_list = NULL)
}
\arguments{
\item{kpi_list}{Id of the KPIs as a list.}

\item{municipality_list}{Id of the municipalities as a list.}

\item{year_list}{The years to get as a list.}
}
\value{
Returns a list that contains data about KPIs in a given municipality by a specific (user-defined) year.
}
\description{
Get Search Results
}
